Add api/v1/schemes.py for API models.
