+++ src/api/v1/schemes.py
@@ -1,0 +1,87 @@
+"""API models."""
+
+
+from uuid import UUID
+from pydantic import BaseModel, Field
+
+
+class FilmShort(BaseModel):
+    """An API model to represent brief film information.
+
+    """
+    id: UUID
+    title: str
+    imdb_rating: float | None
+
+
+class FilmList(BaseModel):
+    """An API model to represent a list of films with a paginator.
+
+    """
+    total: int
+    page: int
+    size: int
+    prev: str | None
+    next: str | None
+    results: list[FilmShort]
+
+
+class GenreInFilm(BaseModel):
+    """An API model to represent genre information within FilmFull class.
+
+    """
+    id: UUID
+    name: str
+
+
+class Person(BaseModel):
+    """An API model to represent person information within FilmFull class.
+
+    """
+    id: UUID
+    name: str | None
+
+
+class FilmFull(FilmShort, BaseModel):
+    """An API model to represent detailed information on a film.
+
+    """
+    description: str | None
+    genres: list[GenreInFilm] | None = Field(default=[])
+    actors: list[Person] | None = Field(default=[])
+    writers: list[Person] | None = Field(default=[])
+    directors: list[Person] | None = Field(default=[])
+
+
+class Genre(BaseModel):
+    id: str
+    name: str
+
+
+class GenreList(BaseModel):
+    results: list[Genre]
+
+
+class FilmPersonRoles(BaseModel):
+    id: str
+    roles: list[str]
+
+
+class PersonShortFilmInfo(BaseModel):
+    id: str
+    title: str
+    imdb_rating: float
+
+
+class PersonShortFilmInfoList(BaseModel):
+    results: list[PersonShortFilmInfo]
+
+
+class Person(BaseModel):
+    id: str
+    full_name: str
+    films: list[FilmPersonRoles]
+
+
+class PersonList(BaseModel):
+    results: list[Person]
